util/apn/requirements.txt
PyYAML
apns==2.0.1
